refactor call_depth to helpers
