+++ api/opentrons/helpers/helpers.py
@@ -1,8 +1,33 @@
 import functools
+import inspect
 import json
 import numbers
 
 from opentrons.util.vector import Vector
+
+
+def call_depth(instance, call_stack):
+    """
+    Returns depth of the penultimate call relative to calls in
+    the stack by methods of this class
+
+    :param instance: object instance of a class
+    :return: int
+    """
+    class_methods = [
+        func_name
+        for func_name, _ in inspect.getmembers(
+            instance, predicate=inspect.ismethod)
+    ]
+    stack_fns = [i.function for i in call_stack]
+
+    # Remove stack methods not in the class
+    calls = filter(lambda m: m in class_methods, stack_fns)
+
+    # Remove private methods
+    calls = filter(lambda x: not x.startswith('_'), calls)
+
+    return len(list(calls)) - 1
 
 
 def is_number(obj):
